Safari-611.1.10/LayoutTests/media/video-canvas-drawing-expected.txt


EVENT(canplaythrough)
EVENT(seeked)
EVENT(seeked)
EVENT(seeked)
EVENT(seeked)
END OF TEST
